back/Chatbot/requirements.txt
# 웹 프레임워크 및 관련 도구
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6

# 인증 및 보안
python-jose>=3.3.0
pyjwt>=2.8.0
python-dotenv>=1.0.0

# OpenAI API
openai>=1.3.0

# 벡터 저장소
chromadb>=0.4.18
langchain>=0.0.267
langchain_community>=0.0.13
langchain-huggingface>=0.0.2
langchain-chroma>=0.0.3

# 임베딩 및 ML 도구
transformers>=4.34.0
sentence-transformers>=2.2.2
torch>=2.1.0

# 기타 유틸리티
tqdm>=4.66.1
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.8.4